Update segment.md (#389)
* Update segment.md

Hey! There's some wacky styling happening here, that I think is triggered by the $ sign perhaps. It's causing an italicized serif to bleed in and seems to start and end with a $ symbol. 

I don't know how to fix this, but wonder if it's related to the brand research and discovery @eerikson had been working on.

* formatting for $ and italics and typo

---------

Co-authored-by: ishamehramixpanel <117322225+ishamehramixpanel@users.noreply.github.com>