Set up spell checking in PRs (#1227)
Use the spell checker we use in the main monorepo. Works pretty good.
Don't enable in PRs yet, just allow people to manually run it on main branch. So we can add obvious false positives to the known word list
When fully deployed this will show inline comments about possible spelling suggestions. This is not configured to block PRs.