Remove mentions to contact support to request higher object lmt  (#1871)
* Remove from object limits info to contact support to request higher limit

Removing that customers with a paid plan can write in the request > 20 object limits. We are only currently only entertaining this for customers who have a plan with $100k+ ARR. "Paid plan" is too broad. Also most customers with $100k+ ARR should have an account team to facilitate this request internally.

* Removing 2nd mention of paid plans can request increase lmt