don't recommend making layout.js "use client" for app router
This is pretty wrong – it's not recommended in nextjs to "use client" for the root layout. Would save people some time not to recommend this pattern and to use a Provider wrapper instead.

In testing, I also found that autotrack doesn't work as expected with app router. We added some additional logic to usePathname and just implement that inside the provider, which I left out for simplicity here but is probably worth flagging as well.